Controllers/Postgres/requirements.txt
flask
sqlalchemy
requests
psycopg2-binary